requirements.txt
fastapi
uvicorn
neo4j
pydantic
python-dotenv
networkx